Finished the last lesson in the ArrayList section
Finished the last lesson in the ArrayList part of Java MasterClass
Section 8: Arrays, Java inbuilt Lists, Autoboxing, and Unboxing. Created
a simple GroceryList class that implements an ArrayList of String
objects to create a grocery list. Implements class methods to provide
data encapsulation, allowing for items to be added, removed, and changed
in list. Also covered some Java built in methods to copy an ArrayList
and to make an Array copy of an ArrayList.