Finished ArrayList challenge part 1
Finished the first ArrayList challenge from Java Master Class.
Implemented a Contacts class to store a name and phone number for a
contact, and a MobilePhone class to initilize an ArrayList of Contacts
objects. The main program then creates a MobilePhone object which in
turn adds, modifies, removes, and prints the contacts accordingly.