Tested the Album and Song classes
got scanner to read from my text file so that I can use that to populate
my list of albums. Was having trouble getting Scanner.hasNext(pattern)
to work as expected. Pattern is a class included int he regex pachage so
I had to figure that out. Then I was using '$' as my token to signify
the end of an album in my text file. However, that token already is
assigned a meaning in regex, it means end of line. So my while loop was
nto working as expected. Once I changed "$" to "@" it worked as
expected.