Final version of Java Master Class Linked List Challenge
Final version of the Java Master Class Linked List challenge in section
8, "Arrays, Java inbuilt Lists, Autoboxing and Unboxing." Implemented a
Song class and and album class, both of which overload equals()
function. The main program creates a file object and sets its filepath
to point to the "Albums_Artists_Songs.txt"file. It then reads all of the
albums from this text file and creates an album object in the ArrayList
of Albums, each of which contains a Linked List of songs on that album.
The program then asks the user to create a playlist of songs and adds
the songs the user specifies. Once finished creating the playlist, the
program plays the playslist, allowing the user to skip ahead of back,
repeat a song, and quit.