Abstract Class project two functions left
I have finished all but the overloaded equals() and set() functions.
Once I have completed those and finished testing, this project will be
finished. Attempted to improve time complexity of remove function by
using the fact that the elements are in order to abort searching for
element to remove if we encounter a value that is less than the one to
be removed. Since the elements are in accending order that would mean
the element is not in the list. Unfortunately I could not figure out how
to cast an object to a generic type. Will be doing further research into
this topic.