Created a package called mylibrary
Was able to create a package called 'mylibrary" that contains a class
called "Series." The series class contains the three math functions
required by the guidelines of this project. I had difficulty getting my
Main.java file to compile and then to run. I kept getting an error that
mylibrary did not exist. I was adding mylibrary to the classpath when
compiling, but not when I was trying to run execute Main. I realized you
must add the path to mylibrary both when you compile and when you
execute any code that imports mylibrary. If I would have permanenlty
added the file path to the classpath this would not have been an issue,
but that is not recommended for individual projects. Also, I was having
trouble specifying the filepath because I am running Git bash on a
Windows machine, so the filepath was not accurate. I had to navigate
through windows explorer, find the filepath, copy that, and then paste
it into the Git bash command prompt, at which point it worked. I also then realized due to the file stucture, I could use the shortcut '..' instead of the full filepath.