Initial commit for section on Java FX
This commit is just for a basic hello world program using the Java FX.
Java FX is a series of module that allow you to write GUIs in Java. For
this program to work I had to download the appropriat SDK and then
create an environment variable to point to path/to/javafx-12.0.1/lib.
After doing this I was able to compile and execute the program from the
MINGW64 command line.